+++ test/t/system.t
@@ -16,9 +16,13 @@ for chroot in chroot fakechroot; do
         echo 'something' > testtree/$chroot-file
         echo "cat /$chroot-file" > testtree/$chroot-system.sh
 
-        t=`( $srcdir/$chroot.sh testtree /bin/test-system 'trap "exit 0" USR1; for i in $(seq 1 10240); do echo $$; done' | while read pid; do readlink /proc/$pid/exe; kill -USR1 $pid; exit; done ) 2>&1`
-        test "$t" = "$fakedir/bin/sh" || not
-        ok "$chroot system shell is" $t
+        if [ "`uname -s`" = "GNU/kFreeBSD" ]; then
+            skip 1 `uname -s`
+        else
+            t=`( $srcdir/$chroot.sh testtree /bin/test-system 'trap "exit 0" USR1; for i in $(seq 1 10240); do echo $$; done' | while read pid; do readlink /proc/$pid/exe; kill -USR1 $pid; exit; done ) 2>&1`
+            test "$t" = "$fakedir/bin/sh" || not
+            ok "$chroot system shell is" $t
+        fi
 
         t=`$srcdir/$chroot.sh testtree /bin/test-system ". /$chroot-system.sh" 2>&1`
         test "$t" = "something" || not
